Resources.resx -------------- Added ConnectionBroken string
MySqlDateTime.cs
----------------
Replaced IComparable implementation 

MySqlDecimal
------------
Now using InvariantCulture for parsing

MySqlDouble
MySqlUInt32
-----------
Minor whitespace cleanups

MySqlInt32
----------
Changed to reading 4 bytes when reading int24 in PS mode

MySqlSingle
-----------
Fixed problem where value was being converted to a double.

MySqlUByte
----------
Fixed problem where value was being converted to an sbyte.

MySqlUInt64
-----------
Fixed problem where Int16 was being incorrectly used to do the parsing.

+++ mysqlclient/Types/MySqlSingle.cs
@@ -82,7 +82,7 @@ namespace MySql.Data.Types
 
 		void IMySqlValue.WriteValue(MySqlStreamWriter writer, bool binary, object val, int length)
 		{
-			double v = Convert.ToSingle(val);
+			Single v = Convert.ToSingle(val);
 			if (binary)
 				writer.Write(BitConverter.GetBytes(v));
 			else
